fix: fix training bug

+++ causvid/models/wan22/wan_base/modules/model.py
@@ -139,13 +139,13 @@ class WanSelfAttention(nn.Module):
             v = self.v(x).view(b, s, n, d)
             return q, k, v
         q, k, v = qkv_fn(x)
-
         x = attention(
             q=rope_apply(q, grid_sizes, freqs),
             k=rope_apply(k, grid_sizes, freqs),
             v=v,
             k_lens=seq_lens,
-            window_size=self.window_size)
+            window_size=self.window_size,
+            dtype=x.dtype)
 
         # x = flash_attention(
         #     q=rope_apply(q, grid_sizes, freqs),
@@ -188,7 +188,7 @@ class WanCrossAttention(WanSelfAttention):
             v = self.v(context).view(b, -1, n, d)
 
         # compute attention
-        x = attention(q, k, v, k_lens=context_lens)
+        x = attention(q, k, v, k_lens=context_lens, dtype=x.dtype)
         # x = flash_attention(q, k, v, k_lens=context_lens)
 
         # output
@@ -255,7 +255,7 @@ class WanAttentionBlock(nn.Module):
         # with torch.amp.autocast('cuda', dtype=torch.float32):
         e = (self.modulation + e).chunk(6, dim=1)
         # assert e[0].dtype == torch.float32
-
+        
         # self-attention
         y = self.self_attn(
             self.norm1(x) * (1 + e[1]) + e[0], seq_lens, grid_sizes, 
@@ -462,7 +462,6 @@ class WanModel(ModelMixin, ConfigMixin):
 
         if y is not None:
             x = [torch.cat([u, v], dim=0) for u, v in zip(x, y)]
-
         # embeddings
         x = [self.patch_embedding(u.unsqueeze(0)) for u in x]
         grid_sizes = torch.stack(
